+++ modules/Shmuelie.Windows/Shmuelie.Windows.psm1
@@ -7,12 +7,12 @@ $exportedCmdlets = @()
 $binaryModule = Join-Path $PSScriptRoot 'bin\Shmuelie.Windows.Cmdlets.dll'
 if (Test-Path $binaryModule) {
     Import-Module $binaryModule -Force -ErrorAction Stop
-    $exportedCmdlets = @('Get-ServiceProcess', 'Get-SubstDrive', 'New-SubstDrive', 'Remove-SubstDrive')
+    $exportedCmdlets = @('Get-InstalledApplications', 'Get-ServiceProcess', 'Get-SubstDrive', 'New-SubstDrive', 'Remove-SubstDrive')
 }
 
 $exportParams = @{
     Function = @(
-        'Get-AppInstallerApp', 'Get-InstalledApplications',
+        'Get-AppInstallerApp',
         'Get-WindowsTerminalSettings', 'Get-WindowsTerminalProfile',
         'Start-WindowsPerformanceRecorder', 'Stop-WindowsPerformanceRecorder',
         'Update-AppInstallerApp'